[LANDING PAGE] Remplace l'ancienne landing par la nouvelle
